requirements.txt
fastapi
uvicorn
transformers
bitsandbytes
python-dotenv
accelerate
huggingface_hub